chore: 페이지 로딩이 되지 않는 문제를 해결하기 위해 임시로 selector 위치 조정
